moved modisco functions to a separate file
